CRS move voip ee files inside a seperate folder
