+++ apps/meteor/ee/app/livechat-enterprise/server/index.js
@@ -35,7 +35,6 @@ import './lib/routing/LoadRotation';
 import './lib/AutoCloseOnHoldScheduler';
 import { onLicense } from '../../license/server';
 import './business-hour';
-import './services/voipService';
 
 onLicense('livechat-enterprise', () => {
 	require('./api');
