+++ apps/meteor/ee/app/voip-enterprise/server/index.ts
@@ -1,0 +1,1 @@
+import './services/voipService';
