+++ apps/meteor/ee/server/index.ts
@@ -7,6 +7,7 @@ import '../app/auditing/server/index';
 import '../app/authorization/server/index';
 import '../app/canned-responses/server/index';
 import '../app/livechat-enterprise/server/index';
+import '../app/voip-enterprise/server/index';
 import '../app/settings/server/index';
 import '../app/teams-mention/server/index';
 import './api';
